Reorganized utility functions
Moved the utility functions into a "utilities" folder and separate files based on their use
